requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.4.1
python-dotenv==1.0.0
